+++ Sources/Scenes/Stories/StoriesViewController.swift
@@ -6,7 +6,7 @@ class StoriesViewController: UIViewController {
 
     private let feed: APIClient.StoryFeed
 
-    private var segmentedProgress: StoriesProgressBar!
+    private var progress: StoriesProgressBar!
 
     private let menuButton: UIButton = {
         let button = UIButton()
@@ -51,12 +51,12 @@ class StoriesViewController: UIViewController {
     override func viewDidLoad() {
         view.backgroundColor = .black
 
-        segmentedProgress = StoriesProgressBar(numberOfSegments: player.player.items().count)
-        segmentedProgress.translatesAutoresizingMaskIntoConstraints = false
-        segmentedProgress.topColor = UIColor.white
-        segmentedProgress.padding = 5.0
-        segmentedProgress.bottomColor = UIColor.white.withAlphaComponent(0.25)
-        segmentedProgress.dataSource = self
+        progress = StoriesProgressBar(numberOfSegments: feed.stories.count)
+        progress.translatesAutoresizingMaskIntoConstraints = false
+        progress.topColor = UIColor.white
+        progress.padding = 5.0
+        progress.bottomColor = UIColor.white.withAlphaComponent(0.25)
+        progress.dataSource = self
 
         let background = UIView()
         background.translatesAutoresizingMaskIntoConstraints = false
@@ -70,7 +70,7 @@ class StoriesViewController: UIViewController {
         content.backgroundColor = .brandColor
         background.addSubview(content)
 
-        content.addSubview(segmentedProgress)
+        content.addSubview(progress)
 
         let buttonStack = UIStackView()
         buttonStack.translatesAutoresizingMaskIntoConstraints = false
@@ -172,10 +172,10 @@ class StoriesViewController: UIViewController {
         ])
 
         NSLayoutConstraint.activate([
-            segmentedProgress.leftAnchor.constraint(equalTo: content.leftAnchor, constant: 20),
-            segmentedProgress.rightAnchor.constraint(equalTo: buttonStack.leftAnchor, constant: -20),
-            segmentedProgress.heightAnchor.constraint(equalToConstant: 4),
-            segmentedProgress.centerYAnchor.constraint(equalTo: buttonStack.centerYAnchor),
+            progress.leftAnchor.constraint(equalTo: content.leftAnchor, constant: 20),
+            progress.rightAnchor.constraint(equalTo: buttonStack.leftAnchor, constant: -20),
+            progress.heightAnchor.constraint(equalToConstant: 4),
+            progress.centerYAnchor.constraint(equalTo: buttonStack.centerYAnchor),
         ])
 
         NSLayoutConstraint.activate([
@@ -209,13 +209,14 @@ class StoriesViewController: UIViewController {
             print("AVAudioSession error: \(error)")
         }
 
-        player.play()
-        segmentedProgress.startAnimation()
+        player.playTrack()
+        progress.startAnimation()
+        progress.isPaused = true
     }
 
     // @TODO allow deselecting reaction?
     @objc private func didReact(_ sender: UIButton) {
-        let item = player.currentItem()
+        let item = feed.stories[player.currentTrack]
 
         if feed.user.id == UserDefaults.standard.integer(forKey: UserDefaultsKeys.userId) {
             return
@@ -237,20 +238,20 @@ class StoriesViewController: UIViewController {
     }
 
     @objc private func exitTapped() {
-        player.stop()
+        player.pause()
         dismiss(animated: true)
     }
 
     @objc private func menuTapped() {
-        let item = player.currentItem()
+        let item = feed.stories[player.currentTrack]
 
         player.pause()
-        segmentedProgress.isPaused = true
+        progress.isPaused = true
 
         let menu = AlertController(title: nil, message: nil, preferredStyle: .actionSheet)
         menu.willDismissHandler = {
             self.player.unpause()
-            self.segmentedProgress.isPaused = false
+            self.progress.isPaused = false
         }
 
         let delete = UIAlertAction(title: NSLocalizedString("delete", comment: ""), style: .destructive, handler: { _ in
@@ -272,12 +273,25 @@ class StoriesViewController: UIViewController {
 }
 
 extension StoriesViewController: StoryPlayerDelegate {
-    func didReachEnd() {
-        player.stop()
-        dismiss(animated: true)
+    func didStartBuffering(_: StoryPlayer) {
+        progress.isPaused = true
     }
 
-    func startedPlaying(story: APIClient.Story) {
+    func didEndBuffering(_: StoryPlayer) {
+        if !progress.isPaused {
+            return
+        }
+
+        progress.isPaused = false
+    }
+
+    func didStartPlaying(_: StoryPlayer, itemAt index: Int) {
+        let story = feed.stories[index]
+
+        if index != 0, progress.currentIndex < index {
+            progress.skip()
+        }
+
         posted.text = Date(timeIntervalSince1970: TimeInterval(story.deviceTimestamp)).timeAgoDisplay()
 
         thumbsUp.count = 0
@@ -297,10 +311,15 @@ extension StoriesViewController: StoryPlayerDelegate {
             }
         }
     }
+
+    func didReachEnd(_ player: StoryPlayer) {
+        player.pause()
+        dismiss(animated: true)
+    }
 }
 
 extension StoriesViewController: StoriesProgressBarDataSource {
     func storiesProgressBar(progressBar _: StoriesProgressBar, durationForItemAt index: Int) -> TimeInterval {
-        return TimeInterval(Float(CMTimeGetSeconds(player.playerItems[index].asset.duration)))
+        return player.duration(for: index)
     }
 }