+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 
  - Allowed skipping in stories.
+ - When stories are buffering, the progress bar now pauses so the story and progress do not look out of sync.
 
 ## [2.15.1] - 2021-03-04